✨ feat: Add cache control options for Shopify API fetch
This commit adds cache control options for the fetch request to the Shopify API in the `getProducts` function. The `cache` property has been commented out to disable caching, and the `next` property has been added to enable automatic cache revalidation every 10 seconds. This change improves the performance and freshness of the data retrieved from the API.